django_crm/requirements.txt
asgiref==3.7.2
Django==4.2.4
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.0
protobuf==4.21.12
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
